Updated Readmes
Made the .txt readme an .rtf to combat bad formatting. Modified both
readmes to remove quotes from "Windows To Go" and changed supported
powershell version to 5.0 instead of 4.0. This fixes the issue we have
where the dism version on Windows 8 wont work with this function.
Version 5.0 ensures Windows 10 or later and the right dism version.

The module and manifest were also updated to check for version 5.0 of
powershell instead of version 4.0.